Allow parsing of decimal values of the Retry_after header (#1355)

+++ modules/org.restlet/src/main/java/org/restlet/engine/header/HeaderUtils.java
@@ -768,7 +768,7 @@ public class HeaderUtils {
                     if (retryAfter == null) {
                         // The date might be expressed as a number of seconds
                         try {
-                            int retryAfterSecs = Integer.parseInt(header.getValue());
+                            int retryAfterSecs = (int) Double.parseDouble(header.getValue());
                             java.util.Calendar calendar = java.util.Calendar.getInstance();
                             calendar.add(java.util.Calendar.SECOND, retryAfterSecs);
                             retryAfter = calendar.getTime();
